conversion of some examples to unit tests, even without testable conditions, if they don't throw an exception it means something sort of works.
other misc changes and cleanup
