+++ pom.xml
@@ -279,19 +279,20 @@
 		    <artifactId>xalan</artifactId>
 		    <version>2.7.2</version>
 		</dependency>
-		<!-- https://mvnrepository.com/artifact/xerces/xerces -->
-		<dependency>
-		    <groupId>xerces</groupId>
-		    <artifactId>xerces</artifactId>
-		    <version>2.4.0</version>
-		</dependency>
+        <!-- https://mvnrepository.com/artifact/xerces/xercesImpl -->
+        <dependency>
+            <groupId>xerces</groupId>
+            <artifactId>xercesImpl</artifactId>
+            <version>2.12.0</version>
+        </dependency>
+
 		<!-- https://mvnrepository.com/artifact/xmlrpc/xmlrpc -->
 		<dependency>
 		    <groupId>xmlrpc</groupId>
 		    <artifactId>xmlrpc</artifactId>
 		    <version>2.0</version>
 		</dependency>
-		
+
