+++ src/test/java/de/unidu/is/pdatalog/examples/StartsWithExample.java
@@ -4,8 +4,8 @@ package de.unidu.is.pdatalog.examples;
 import de.unidu.is.pdatalog.EDBRelation;
 import de.unidu.is.pdatalog.RelationBase;
 import de.unidu.is.pdatalog.parser.Parser;
-import de.unidu.is.util.DB;
 import de.unidu.is.util.HSQLDBEmbeddedDB;
+import junit.framework.TestCase;
 
 /**
  * A simple example.
@@ -14,11 +14,10 @@ import de.unidu.is.util.HSQLDBEmbeddedDB;
  * @version $Revision: 1.2 $, $Date: 2005/03/18 09:54:49 $
  * @since 19-Oct-2004
  */
-public class StartsWithExample {
+public class StartsWithExample extends TestCase {
 
-    public static void main(String[] args) {
-        DB db = new HSQLDBEmbeddedDB();
-        RelationBase base = new RelationBase(db);
+    public void test1() {
+        RelationBase base = new RelationBase(new HSQLDBEmbeddedDB());
 
         new EDBRelation(base, "s", 1, true);
         base.add(Parser.parseFact("s(abc)."));
@@ -26,6 +25,8 @@ public class StartsWithExample {
         base.add(Parser.parseFact("s(xabc)."));
         base.add(Parser.parseFact("s(abd)."));
 
-        System.out.println(base.query("?- s(X) & startswith(X,ab)."));
+        /* TODO check if this is actual correct result */
+        assertEquals("[('ab') | (1.0*PROB)., ('abc') | (1.0*PROB)., ('abd') | (1.0*PROB).]",
+                base.query("?- s(X) & startswith(X,ab).").toString());
     }
 }
