+++ src/test/java/de/unidu/is/text/AbstractFilterTest.java
@@ -15,7 +15,7 @@ specific language governing permissions and limitations under the License.
 
 
 // $Id: AbstractFilterTest.java,v 1.5 2005/02/21 17:29:29 huesselbeck Exp $
-package test.de.unidu.is.text;
+package de.unidu.is.text;
 
 import de.unidu.is.text.AbstractFilter;
 import de.unidu.is.util.CollectionUtilities;
